update gaia README.md
