--- .env
@@ -1,29 +1,0 @@
-# Development settings
-GOOGLE_API_KEY="AIzaSyCvnJqo67NdiIWsXzx7QGSc5kjyf7CtqTk"
-GOOGLE_SEARCH_ENGINE_ID="038388dcda86e4e0a"
-
-LLM_API_KEY="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"
-LLM_BASE_URL="https://openrouter.ai/api/v1"
-LLM_MODEL_NAME="anthropic/claude-3.7-sonnet"
-
-E2B_API_KEY="e2b_425d6730de1d9f1fab0afabac373dca6f829d196"
-
-FILESYSTEM_SERVER_WORKDIR="/Users/yuchengyue/AWorld/gaia/gaia_dataset"
-
-GAIA_DATASET_PATH="/Users/yuchengyue/AWorld/gaia/gaia_dataset"
-
-LOG_FILE_PATH="run_super_agent.log"
-
-AUDIO_LLM_API_KEY="sk-zk21f407a45d74d961f600db8866504c69067e3a55903686"
-AUDIO_LLM_BASE_URL="https://openrouter.ai/api/v1"
-AUDIO_LLM_MODEL_NAME="anthropic/claude-3.7-sonnet"
-
-IMAGE_LLM_API_KEY="sk-WzTuFmfF2pgbzWfflwwMcVbhlMiJOCnJrjHdXWUY2cipRnoY"
-IMAGE_LLM_BASE_URL="https://api.bianxie.ai/v1"
-IMAGE_LLM_MODEL_NAME="gpt-4.1"
-
-VIDEO_LLM_API_KEY="sk-zk21f407a45d74d961f600db8866504c69067e3a55903686"
-VIDEO_LLM_BASE_URL="https://openrouter.ai/api/v1"
-VIDEO_LLM_MODEL_NAME="anthropic/claude-3.7-sonnet"
-
-APIFY_API_TOKEN="apify_api_Qd01Oz3wcdxeVhmmpZbXncmC4DGbkA2TkB93"
