+++ aworld/config/mcp.json
@@ -1,16 +1,69 @@
 {
   "mcpServers": {
-    "amap-amap-sse": {
-      "url": "https://mcp.amap.com/sse?key=YOUR_API_KEY",
-      "timeout": 5.0,
-      "sse_read_timeout": 300.0
-    },
-    "tavily-mcp": {
-      "command": "npx",
-      "args": ["-y", "tavily-mcp@0.1.2"],
-      "env": {
-        "TAVILY_API_KEY": "YOUR_API_KEY"
+      "e2b-server": {
+          "command": "npx",
+          "args": ["-y", "@e2b/mcp-server"],
+          "env": { "E2B_API_KEY": "${E2B_API_KEY}" }
+      },
+      "playwright": {
+          "command": "npx",
+          "args": ["-y", "@executeautomation/playwright-mcp-server"]
+      },
+      "filesystem": {
+          "command": "npx",
+          "args": [
+              "-y",
+              "@modelcontextprotocol/server-filesystem",
+              "${FILESYSTEM_SERVER_WORKDIR}"
+          ]
+      },
+      "terminal-controller": {
+          "command": "python",
+          "args": ["-m", "terminal_controller"]
+      },
+      "calculator": {
+          "command": "python",
+          "args": ["-m", "mcp_server_calculator"]
+      },
+      "excel": {
+          "command": "uv",
+          "args": ["run", "excel-mcp-server"]
+      },
+      "google-search": {
+          "command": "npx",
+          "args": [
+              "-y",
+              "@adenot/mcp-google-search"
+          ],
+          "env": {
+              "GOOGLE_API_KEY": "${GOOGLE_API_KEY}",
+              "GOOGLE_SEARCH_ENGINE_ID": "${GOOGLE_SEARCH_ENGINE_ID}"
+          }
+      },
+      "ms-playwright": {
+          "command": "npx",
+          "args": [
+            "@playwright/mcp@latest"
+          ],
+          "env": {
+            "PLAYWRIGHT_TIMEOUT": "120000"
+          }
+      },
+      "audio_server": {
+          "command": "python",
+          "args": ["-m", "mcp_servers.audio_server"]
+      },
+      "image_server": {
+          "command": "python",
+          "args": ["-m", "mcp_servers.image_server"]
+      },
+      "youtube_download_server": {
+          "command": "python",
+          "args": ["-m", "mcp_servers.youtube_download_server"]
+      },
+      "video_server": {
+          "command": "python",
+          "args": ["-m", "mcp_servers.video_server"]
       }
-    }
   }
 }
