--- mcp.json
@@ -1,69 +1,0 @@
-{
-    "mcpServers": {
-        "e2b-server": {
-            "command": "npx",
-            "args": ["-y", "@e2b/mcp-server"],
-            "env": { "E2B_API_KEY": "${E2B_API_KEY}" }
-        },
-        "playwright": {
-            "command": "npx",
-            "args": ["-y", "@executeautomation/playwright-mcp-server"]
-        },
-        "filesystem": {
-            "command": "npx",
-            "args": [
-                "-y",
-                "@modelcontextprotocol/server-filesystem",
-                "${FILESYSTEM_SERVER_WORKDIR}"
-            ]
-        },
-        "terminal-controller": {
-            "command": "python",
-            "args": ["-m", "terminal_controller"]
-        },
-        "calculator": {
-            "command": "python",
-            "args": ["-m", "mcp_server_calculator"]
-        },
-        "excel": {
-            "command": "uv",
-            "args": ["run", "excel-mcp-server"]
-        },
-        "google-search": {
-            "command": "npx",
-            "args": [
-                "-y",
-                "@adenot/mcp-google-search"
-            ],
-            "env": {
-                "GOOGLE_API_KEY": "${GOOGLE_API_KEY}",
-                "GOOGLE_SEARCH_ENGINE_ID": "${GOOGLE_SEARCH_ENGINE_ID}"
-            }
-        },
-        "ms-playwright": {
-            "command": "npx",
-            "args": [
-              "@playwright/mcp@latest"
-            ],
-            "env": {
-              "PLAYWRIGHT_TIMEOUT": "120000"
-            }
-        },
-        "audio_server": {
-            "command": "python",
-            "args": ["-m", "mcp_servers.audio_server"]
-        },
-        "image_server": {
-            "command": "python",
-            "args": ["-m", "mcp_servers.image_server"]
-        },
-        "youtube_download_server": {
-            "command": "python",
-            "args": ["-m", "mcp_servers.youtube_download_server"]
-        },
-        "video_server": {
-            "command": "python",
-            "args": ["-m", "mcp_servers.video_server"]
-        }
-    }
-}
